requirements.txt
flask
flask-sqlalchemy
bcrypt
python-dotenv
Flask-Bcrypt 
FLask-Migrate 
Flask-SQLAlchemy 
